Trinity.C: Network: refactoring.
