+++ src/Trinity.C/Trinity.C.Linux.vcxproj
@@ -157,7 +157,6 @@
     <ClCompile Include="src\Network\Client\windows\ClientSocket.cpp" />
     <ClCompile Include="src\Network\network_exported_methods.cpp" />
     <ClCompile Include="src\Network\ProtocolConstants.cpp" />
-    <ClCompile Include="src\Network\Server\Common.cpp" />
     <ClCompile Include="src\Network\Server\iocp\TrinitySocketServer.ContextObject.cpp" />
     <ClCompile Include="src\Network\Server\iocp\TrinitySocketServer.cpp" />
     <ClCompile Include="src\Network\Server\iocp\TrinitySocketServer.Recv.cpp" />
@@ -166,6 +165,7 @@
     <ClCompile Include="src\Network\Server\posix\epoll.cpp" />
     <ClCompile Include="src\Network\Server\posix\kqueue.cpp" />
     <ClCompile Include="src\Network\Server\posix\TrinitySocketServer.cpp" />
+    <ClCompile Include="src\Network\Server\TrinityServer.cpp" />
     <ClCompile Include="src\Network\Server\TrinitySocketServerTest.cpp" />
     <ClCompile Include="src\Runtime\dllmain.cpp" />
     <ClCompile Include="src\Runtime\init.cpp" />
