+++ src/Trinity.C/Trinity.C.vcxproj
@@ -345,7 +345,7 @@ _SCL_SECURE_NO_WARNINGS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
     <ClCompile Include="src\Network\Client\windows\ClientSocket.cpp" />
     <ClCompile Include="src\Network\network_exported_methods.cpp" />
     <ClCompile Include="src\Network\ProtocolConstants.cpp" />
-    <ClCompile Include="src\Network\Server\Common.cpp" />
+    <ClCompile Include="src\Network\Server\TrinityServer.cpp" />
     <ClCompile Include="src\Network\Server\iocp\TrinitySocketServer.ContextObject.cpp" />
     <ClCompile Include="src\Network\Server\iocp\TrinitySocketServer.cpp" />
     <ClCompile Include="src\Network\Server\iocp\TrinitySocketServer.Recv.cpp" />
