+++ src/Trinity.C/Trinity.C.vcxproj.filters
@@ -338,9 +338,6 @@
     <ClCompile Include="include\os\platforms\posix.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="src\Network\Server\Common.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="src\Network\Server\TrinitySocketServerTest.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -351,6 +348,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="src\Network\ProtocolConstants.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="src\Network\Server\TrinityServer.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
