+++ src/Trinity.C/src/Network/Server/iocp/TrinitySocketServer.h
@@ -2,6 +2,8 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE.md file in the project root for full license information.
 //
+#include "os/os.h"
+#ifdef TRINITY_PLATFORM_WINDOWS
 #include "Network/Server/iocp/Common.h"
 #include "Network/Server/TrinityServer.h"
 #include <Threading/TrinitySpinlock.h>
@@ -39,3 +41,4 @@ namespace Trinity
         bool ProcessIOCompletion(void* pContext, uint32_t opType, uint32_t bytesTransferred);
     }
 }
+#endif
